Shift create report to a separate file
